crawler/requirements.txt
Scrapy~=2.8.0
psycopg2-binary~=2.9.6